Implementation des tags, revu des tables pivots et relations, restructure des resources des apis

+++ app/Http/Controllers/CommentController.php
@@ -15,7 +15,7 @@ class CommentController extends Controller
         try {
             $comments = CommentResource::collection(Comment::all());
         return response()->json([
-            'Comment' => $comments,
+            'data' => $comments
         ], 200);
         } catch (\Exception $message) {
             return response()->json([
@@ -28,19 +28,16 @@ class CommentController extends Controller
     public function store(CommentRequest $request)
     {
         try {
-          
             $comment = Comment::create([
-                "user_id" => auth()->user()->id, 
+                "user_id" => auth()->user()->id,
                 "post_id" => $request['post_id'],
-                "comment" => $request['comment'],
-                "user_name" => auth()->user()->name,
+                "comment" => $request['comment']
             ]);
-    
+
             return response()->json([
-                'Message' => "Commentaire ajouté avec success",
-                'Comment' => $comment ,
+                'data' => $comment ,
             ], 200);
-    
+
         } catch (\Exception $message) {
             return response()->json([
                 'Erreur : ' => $message->getMessage(),
@@ -56,23 +53,27 @@ class CommentController extends Controller
         } catch (\Exception $message) {
             return response()->json(["Erreur :" => $message->getMessage()], 403);
         }
-        
+
     }
 
-  
+
     public function update(CommentRequest $request, Comment $comment)
     {
         try {
-          
+            if (auth()->user()->id !==  $comment->user_id){
+                return response()->json([
+                    'Erreur :' => 'Seul l utilisateur peut modifier ce commentaire'
+                ]);
+            }
             $comment->update([
                 "comment" => $request['comment']
             ]);
-    
+
             return response()->json([
                 'Message' => "Commentaire modfié avec success",
                 'Post modifié' => $comment , //->id,
             ], 200);
-    
+
         } catch (\Exception $message) {
             return response()->json([
                 'Erreur : ' => $message->getMessage(),
@@ -83,6 +84,11 @@ class CommentController extends Controller
     public function destroy(Comment $comment)
     {
         try {
+            if (auth()->user()->id !==  $comment->user_id){
+                return response()->json([
+                    'Erreur :' => 'Seul l utilisateur peut supprimer ce commentaire'
+                ]);
+            }
             $comment->delete();
             return response()->json(["Message" => "Commentaire supprimé"], 200);
         } catch (\Exception $message) {
@@ -90,18 +96,16 @@ class CommentController extends Controller
                 'Erreur : ' => $message->getMessage(),
             ], 403);
         }
-        
+
     }
 
     public function commentOfOneUser(){
         try {
             // Récupérer uniquement les articles de l'utilisateur connecté
             $post_all = CommentResource::collection(Comment::where('user_id', auth()->id())->get());
-    
-            return response()->json([
-                'Comments' => $post_all,
-            ]);
-    
+
+            return response()->json($post_all);
+
         } catch (\Exception $message) {
             return response()->json([
                 'Erreur' => $message->getMessage(),
